refactor: move exclude objects components in status directory
