+++ src/components/panels/Status/ExcludeObjectDialogMap.vue
